rearrange teardowns (#80)
* made teardowns executable

* full teardown script

* Create teardown

* made executable

* Update teardown

* remove pipefail

* Revert "remove pipefail from teardowns"

* check if file exists otherwise ignore

* rearrange teardowns added remove answers to teardowns

---------

+++ labs/drill/application-monitoring/teardown.sh
@@ -3,5 +3,11 @@ set -euo pipefail
 kubectl delete --ignore-not-found -f https://github.com/kubernetes-sigs/metrics-server/releases/latest/download/components.yaml
 kubectl delete --ignore-not-found -f ~/mycode/yaml/ctce-drill-monitoring.yaml
 sleep 60
-rm ~/cpu-consume.txt
- "Teardown complete"
+if [ -e ~/cpu-consume.txt ] 
+then
+  echo "deleting"
+  rm ~/cpu-consume.txt
+else
+  echo "file does not exist"
+fi
+echo "Teardown complete"
