+++ labs/drill/storage/teardown.sh
@@ -1,6 +1,26 @@
 #!/bin/bash
 set -euo pipefail
-kubectl delete --ignore-not-found -f ~/alta3-pv.yaml
-kubectl delete --ignore-not-found -f ~/nginx-pvc.yaml
-kubectl delete --ignore-not-found -f ~/nginx-with-pv.yaml
+if [ -e ~/nginx-pvc.yaml ] 
+then
+  echo "deleting"
+  kubectl delete --ignore-not-found -f ~/nginx-pvc.yaml
+else
+  echo "file does not exist"
+fi
+if [ -e ~/alta3-pv.yaml ] 
+then
+  echo "deleting"
+  kubectl delete --ignore-not-found -f ~/alta3-pv.yaml
+else
+  echo "file does not exist"
+fi
+if [ -e ~/nginx-with-pv.yaml ] 
+then
+  echo "deleting"
+  kubectl delete --ignore-not-found -f ~/nginx-with-pv.yaml
+else
+  echo "file does not exist"
+fi
+kubectl delete --ignore-not-found -f ~/mycode/yaml/ctce-answers-storage-nginx-pvc.yaml
+kubectl delete --ignore-not-found -f ~/mycode/yaml/ctce-answers-storage-nginx-pv.yaml
 echo "Teardown complete"
